Changed Unicorn capistrano setup

+++ config/deploy.rb
@@ -87,7 +87,11 @@ namespace :deploy do
   after "deploy", "deploy:restart_unicorn"
 
   task :restart_unicorn, roles: :app do
-    run "cd #{current_path} && bundle install --binstubs"
-    sudo "#{current_path}/bin/unicorn -D -c #{current_path}/config/unicorn.rb -E production"
+    if File.exist? '#{current_path}/bin/unicorn'
+      run "#{current_path}/bin/unicorn -D -c #{current_path}/config/unicorn.rb -E production"
+    else
+      run "cd #{current_path} && bundle install --binstubs"
+      run "#{current_path}/bin/unicorn -D -c #{current_path}/config/unicorn.rb -E production"
+    end
   end
 end
